refactor(client): [CLI-026] :recycle: Use String Format function
